DRAFT Compiler FE introduce souschef
on-going draft of souschef; common codes for circlechef and tflchef

ONE-DCO-1.0-ONE-DCO-1.0-ONE-DCO-1.0-Signed-off-by: SaeHie Park <saehie.park@gmail.com>
