+++ compiler/circlechef/core/CMakeLists.txt
@@ -6,3 +6,4 @@ target_include_directories(circlechef_core PRIVATE src)
 target_link_libraries(circlechef_core circlechef_proto)
 target_link_libraries(circlechef_core circlechef_log)
 target_link_libraries(circlechef_core mio_circle)
+target_link_libraries(circlechef_core souschef)
