+++ compiler/circlechef/requires.cmake
@@ -5,3 +5,4 @@ require("safemain")
 require("hermes")
 require("hermes-std")
 require("foder")
+require("souschef")
