+++ compiler/souschef/README.md
@@ -1,0 +1,3 @@
+# souschef
+
+_souschef_ is common library for _tflchef_ and _circlechef_
