+++ compiler/tflchef/core/CMakeLists.txt
@@ -6,3 +6,4 @@ target_include_directories(tflchef_core PRIVATE src)
 target_link_libraries(tflchef_core tflchef_proto)
 target_link_libraries(tflchef_core tflchef_log)
 target_link_libraries(tflchef_core mio_tflite)
+target_link_libraries(tflchef_core souschef)
